requirements.txt
Django==2.2
django-crispy-forms==1.10.0
django-tinymce==3.2.0
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
